requirements.txt
django==6.0.2
python-dateutil
